Update raeog.php
From: http://php.net/manual/en/function.curl-setopt.php

CURLOPT_PROGRESSFUNCTION	
A callback accepting five parameters. The first is the cURL resource, the second is the total number of bytes expected to be downloaded in this transfer, the third is the number of bytes downloaded so far, the fourth is the total number of bytes expected to be uploaded in this transfer, and the fifth is the number of bytes uploaded so far.

Note:
The callback is only called when the CURLOPT_NOPROGRESS option is set to FALSE.
Return a non-zero value to abort the transfer. In which case, the transfer will set a CURLE_ABORTED_BY_CALLBACK error.